requirements.txt
ftfy==6.1.3
regex==2023.12.25
tqdm==4.66.1
numpy==1.20.3
pillow==10.0.1
pickleshare==0.7.5
scikit-learn==1.0
git+https://github.com/openai/CLIP.git



